x509check.py
A python-based script (to be run as a cron job) that checks the
expriation date of an X.509 certificate and compare its expiration date
against today's date. If the expiration date is at T-minus X days out
(or if the certificate has already expired), then send an email
notification to a list of recipients warning them about the impending
expiration of the certificate.